satellite-detector/requirements.txt
apscheduler
pythonping
ping3
dnspython
pandas
seaborn
matplotlib
scipy
